toxmm-rewrite: split off main::storage to utils::storage
